Moved panel editing components to it's own folder
