+++ public/app/features/dashboard/panel_editor/EditorTabBody.tsx
