+++ public/app/features/dashboard/panel_editor/GeneralTab.tsx
