+++ public/app/features/dashboard/panel_editor/PanelEditor.tsx
