+++ public/app/features/dashboard/panel_editor/QueriesTab.tsx
