+++ public/app/features/dashboard/panel_editor/QueryInspector.tsx
