+++ public/app/features/dashboard/panel_editor/VisualizationTab.tsx
