+++ public/app/features/dashboard/panel_editor/VizTypePickerPlugin.tsx
